Fix regressions from recent refactoring in tests and switch platform
- Updated `tests/discovery/test_service.py` to use correct imports and mock types.
- Updated `tests/sensor/device/test_meraki_firmware_status.py` to use correct coordinator class.
- Updated `tests/switch/test_mt40_power_outlet.py` to fix imports and mocking of `MerakiAPIClient`.
- Updated `tests/sensor/test_uplink_performance.py` to use correct metric key (`latency`).
- Updated `tests/core/test_api_utils.py` to patch the correct logger location.
- Updated `tests/sensor/test_setup_mt_sensors.py` to fix entity count expectations for MT15 and mock `entity.platform` for translations.
- Updated `custom_components/meraki_ha/sensor/device/switch_port.py` and `custom_components/meraki_ha/switch/switch_port.py` to use `device.switch_ports` instead of `device.ports_statuses`.
- Updated `tests/switch/test_meraki_switch_port_switch.py` to properly mock `coordinator.get_device`.

Verified all 347 tests pass.
