+++ tests/discovery/test_service.py
@@ -7,8 +7,8 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 
 from custom_components.meraki_ha.coordinator import MerakiDataUpdateCoordinator
+from custom_components.meraki_ha.core.api.client import MerakiAPIClient as MerakiClient
 from custom_components.meraki_ha.discovery.service import DeviceDiscoveryService
-from custom_components.meraki_ha.meraki_client import MerakiClient
 from custom_components.meraki_ha.services.camera_service import CameraService
 from custom_components.meraki_ha.services.device_control_service import (
     DeviceControlService,
@@ -23,7 +23,7 @@ from tests.const import MOCK_DEVICE
 @pytest.fixture
 def mock_coordinator_with_devices(
     mock_coordinator: MagicMock,
-) -> MagicMock[MerakiDataUpdateCoordinator]:
+) -> MagicMock:
     """Fixture for a mocked MerakiDataUpdateCoordinator with various devices."""
     wireless_device = replace(MOCK_DEVICE, model="MR36")
     camera_device = replace(MOCK_DEVICE, serial="camera_serial", model="MV12")
@@ -39,13 +39,13 @@ def mock_coordinator_with_devices(
 
 
 @pytest.fixture
-def mock_camera_service() -> AsyncMock[CameraService]:
+def mock_camera_service() -> AsyncMock:
     """Fixture for a mocked CameraService."""
     return AsyncMock()
 
 
 @pytest.fixture
-def mock_control_service() -> MagicMock[DeviceControlService]:
+def mock_control_service() -> MagicMock:
     """Fixture for a mock DeviceControlService."""
     return MagicMock()
 
@@ -57,8 +57,8 @@ def test_discovery_service_init(
     mock_control_service: DeviceControlService,
 ) -> None:
     """Test the initialization of the DeviceDiscoveryService."""
-    mock_meraki_client: MagicMock[MerakiClient] = MagicMock()
-    mock_network_control_service: MagicMock[NetworkControlService] = MagicMock()
+    mock_meraki_client: MagicMock = MagicMock()
+    mock_network_control_service: MagicMock = MagicMock()
 
     service: DeviceDiscoveryService = DeviceDiscoveryService(
         coordinator=mock_coordinator_with_devices,
@@ -115,8 +115,8 @@ async def test_discover_entities_delegates_to_handler(
         mock_wireless_handler_instance.discover_entities.side_effect = mock_aiter_empty
         MockWirelessHandler.return_value = mock_wireless_handler_instance
 
-        mock_meraki_client: MagicMock[MerakiClient] = MagicMock()
-        mock_network_control_service: MagicMock[NetworkControlService] = MagicMock()
+        mock_meraki_client: MagicMock = MagicMock()
+        mock_network_control_service: MagicMock = MagicMock()
 
         service: DeviceDiscoveryService = DeviceDiscoveryService(
             coordinator=mock_coordinator_with_devices,
